Added a temporary version pointing to vds.icefilms.info just in case you wanna test something...  I don't know what but whatever, I needed it. To use just install vds.icefilms.apk.  It's version 1.1 so it'll overwrite the last signed one I put up.  Then when icefilms.info fixes its new site just install www.icefilms.apk.  It's at v1.2 so it'll just "update" the v1.1 vds one. (or so I'm assuming)
Took out the test device since I've decided I'm just not gonna keep developing this project.  I'm gonna instead create a replacement made of the shell of this with entirely new downloading code using a google app engine app to offload the parsing, in addition to caching all old info for much less overhead and much more simplicity (and much less mobile data usage, not that you should be downloading high quality divx videos on your mobile plan anyway unless it's unlimited)

So yeah, keep an eye on icequickstream.appspot.com for updates.  For now it's just a way to access/try out the current site (which you have to dig through the page source on a video page of icefilms to get the source ID for anyway...).  Eventually I'll make an actual landing page for info about both the android app and maybe the online service (hopefully I'll actually learn what I'm doing and develop a real web interface with GWT or something for when icefilms is down).